fix: convert component styles to sass, refactor media queries
