+++ components/Container.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from "react";
-import styles from "./Container.module.css";
+import styles from "./Container.module.scss";
 
 interface ContainerProps {
   children?: ReactNode;
