+++ components/Intro.module.scss
@@ -1,3 +1,4 @@
+@import "../styles/breakpoints.scss";
 
 .root {
   position: relative;
@@ -10,16 +11,14 @@
   max-width: var(--layout-content-max-width);
   margin: 0 auto;
 
-  @media (--media-min-small) {
-    background: red;
-    /* background: var(--color-pink); */
+  @media (min-width: $media-min-small) {
     line-height: 1.25;
     padding: 5rem var(--layout-gutter-width-desktop) 0;
   }
-  @media (--media-min-large) {
+  @media (min-width: $media-min-large) {
     padding-top: 8rem;
   }
-  & h2 {
+  h2 {
     line-height: 1;
     margin: 0 0 2rem;
     font-family: var(--font-family-serif);
@@ -28,31 +27,31 @@
     /* @apply --effect-fadeInUp; */
     animation-duration: var(--animation-timing-duration);
     animation-delay: var(--animation-timing-first);
-    @media (--media-min-extra-small) {
+    @media (min-width: $media-min-extra-small) {
       font-size: 4.8rem;
       margin: 0 0 3rem;
     }
-    @media (--media-min-large) {
+    @media (min-width: $media-min-large) {
       font-size: 6.5rem;
       margin: 0 0 4.8rem;
     }
   }
-  & p {
+  p {
     margin: 0 0 2rem;
     /* @apply --effect-fadeIn; */
     animation-duration: var(--animation-timing-duration);
     animation-delay: var(--animation-timing-second);
-    @media (--media-min-small) {
+    @media (min-width: $media-min-small) {
       margin: 0 0 2.5rem;
     }
   }
-  & p:last-child {
+  p:last-child {
     margin-bottom: 0;
   }
 }
 .iconWrapper {
   font-size: 4rem;
-  @media (--media-min-small) {
+  @media (min-width: $media-min-small) {
     font-size: 7rem;
   }
 }
