+++ components/PortfolioItem.module.scss
@@ -1,7 +1,9 @@
 
+@import "../styles/breakpoints.scss";
+
 .root {
   margin-bottom: 4.5rem;
-  @media (--media-min-small) {
+  @media (min-width: $media-min-small) {
     margin-bottom: 5.5rem;
   }
   &:last-child {
@@ -13,11 +15,11 @@
   margin: 0 0 1.7rem;
   font-family: var(--font-family-serif);
   font-size: var(--font-large-size);
-  @media (--media-min-small) {
+  @media (min-width: $media-min-small) {
     margin: 0 0 1.9rem;
     font-size: var(--font-huge-size);
   }
-  @media (--media-min-large) {
+  @media (min-width: $media-min-large) {
     margin: 0 0 2.3rem;
     font-size: var(--font-massive-size);
   }
@@ -39,7 +41,7 @@
       animation-name: bounce-small;
       animation-iteration-count: infinite;
       animation-direction: alternate;
-      @media (--media-min-small) {
+      @media (min-width: $media-min-small) {
         animation-name: bounce;
       }
     }
@@ -59,7 +61,7 @@
   transition: all 0.35s ease;
   font-size: 3rem;
   margin-left: 1.4rem;
-  @media (--media-min-small) {
+  @media (min-width: $media-min-small) {
     font-size: 3.6rem;
     margin-left: 1.9rem;
   }
