+++ components/PortfolioItem.tsx
@@ -1,6 +1,6 @@
 import { cn } from "../lib/helpers";
-import markdownStyles from "./markdown-styles.module.css";
-import styles from "./PortfolioItem.module.css";
+import markdownStyles from "./markdown-styles.module.scss";
+import styles from "./PortfolioItem.module.scss";
 import icons from "../styles/icons.module.scss";
 
 interface Props {
