--- styles/custom-media.scss
@@ -1,8 +1,0 @@
-@custom-media --media-min-extra-small (min-width: 350px);
-@custom-media --media-max-extra-small (max-width: 349px);
-@custom-media --media-min-small (min-width: 768px);
-@custom-media --media-max-small (max-width: 767px);
-@custom-media --media-min-medium (min-width: 1265px);
-@custom-media --media-max-medium (max-width: 1264px);
-@custom-media --media-min-large (min-width: 1600px);
-@custom-media --media-max-large (max-width: 1599px);
