+++ styles/effects.scss
@@ -1,5 +1,3 @@
-// @import "./custom-properties.scss";
-
 :root {
   --solid-underline: {
     background-image: linear-gradient(
